[#163] Enhance user serializer with registration logic and move food request serializers to dedicated package
Subject: Restructure user and request serializers for enhanced functionality

- Moved the `FoodRequestCreateSerializer` and `FoodRequestUpdateSerializer` to a new `request` package to better organize the serializers
- Added a new `RegisterSerializer` that extends the base `UserSerializer` and generates a random password for new users
- Refactored the `UserSerializer` to handle password hashing and generation of a random password for new users
- These changes improve the structure and functionality of the user and request serializer modules, making the codebase more maintainable and extensible
