+++ apps/api/afbcore/serializers/__init__.py
@@ -1,5 +1,6 @@
-from .food_request_serializer import (
-    FoodRequestCreateSerializer,  # noqa
-    FoodRequestUpdateSerializer,
+from .request.food_request_serializer import (
+    FoodRequestCreateSerializer,  # noqa: F401
+    FoodRequestUpdateSerializer,  # noqa: F401
 )
-from .user_serializer import UserSerializer  # noqa
+from .user.register_serializer import RegisterSerializer  # noqa: F401
+from .user.user_serializer import UserSerializer  # noqa: F401
